Splitting rules into separate files
