+++ src/sqlfluff/core/rules/__init__.py
@@ -1,7 +1,18 @@
-"""init py for the new rules crawlers."""
+"""Register all the rule classes with their corresponding rulesets (just std currently)."""
 
-from .std import std_rule_set
+import inspect
+
 from .base import RuleSet, rules_logger  # noqa
+from .config_info import STANDARD_CONFIG_INFO_DICT
+from sqlfluff.core.rules import std
+
+
+std_rule_set = RuleSet(name="standard", config_info=STANDARD_CONFIG_INFO_DICT)
+
+for name, val in std.__dict__.items():
+    # Filter on the item being a rule
+    if name.startswith("Rule_L") and inspect.isclass(val):
+        std_rule_set.register(val)
 
 
 def get_ruleset(name: str = "standard") -> RuleSet:
