+++ src/sqlfluff/core/rules/std/L009.py
@@ -1,0 +1,43 @@
+"""Implementation of Rule L009."""
+
+from sqlfluff.core.rules.base import BaseCrawler, LintResult, LintFix
+from sqlfluff.core.rules.doc_decorators import document_fix_compatible
+
+
+@document_fix_compatible
+class Rule_L009(BaseCrawler):
+    """Files must end with a trailing newline."""
+
+    def _eval(self, segment, siblings_post, parent_stack, **kwargs):
+        """Files must end with a trailing newline.
+
+        We only care about the segment and the siblings which come after it
+        for this rule, we discard the others into the kwargs argument.
+
+        """
+        if len(self.filter_meta(siblings_post)) > 0:
+            # This can only fail on the last segment
+            return None
+        elif len(segment.segments) > 0:
+            # This can only fail on the last base segment
+            return None
+        elif segment.name == "newline":
+            # If this is the last segment, and it's a newline then we're good
+            return None
+        elif segment.is_meta:
+            # We can't fail on a meta segment
+            return None
+        else:
+            # so this looks like the end of the file, but we
+            # need to check that each parent segment is also the last
+            file_len = len(parent_stack[0].raw)
+            pos = segment.pos_marker.char_pos
+            # Does the length of the file, equal the length of the segment plus its position
+            if file_len != pos + len(segment.raw):
+                return None
+
+        ins = self.make_newline(pos_marker=segment.pos_marker.advance_by(segment.raw))
+        # We're going to make an edit because otherwise we would never get a match!
+        return LintResult(
+            anchor=segment, fixes=[LintFix("edit", segment, [segment, ins])]
+        )
