+++ src/sqlfluff/core/rules/std/L027.py
@@ -1,0 +1,57 @@
+"""Implementation of Rule L027."""
+
+from sqlfluff.core.rules.base import LintResult
+from sqlfluff.core.rules.std.L025 import Rule_L025
+
+
+class Rule_L027(Rule_L025):
+    """References should be qualified if select has more than one referenced table/view.
+
+    NB: Except if they're present in a USING clause.
+
+    | **Anti-pattern**
+    | In this example, the reference 'vee' has not been declared
+    | and the variables 'a' and 'b' are potentially ambiguous.
+
+    .. code-block:: sql
+
+        SELECT a, b
+        FROM foo
+        LEFT JOIN vee ON vee.a = foo.a
+
+    | **Best practice**
+    |  Add the references.
+
+    .. code-block:: sql
+
+        SELECT foo.a, vee.b
+        FROM foo
+        LEFT JOIN vee ON vee.a = foo.a
+    """
+
+    def _lint_references_and_aliases(
+        self, aliases, references, col_aliases, using_cols, parent_select
+    ):
+        # Do we have more than one? If so, all references should be qualified.
+        if len(aliases) <= 1:
+            return None
+        # A buffer to keep any violations.
+        violation_buff = []
+        # Check all the references that we have.
+        for r in references:
+            this_ref_type = r.qualification()
+            if (
+                this_ref_type == "unqualified"
+                and r.raw not in col_aliases
+                and r.raw not in using_cols
+            ):
+                violation_buff.append(
+                    LintResult(
+                        anchor=r,
+                        description="Unqualified reference {0!r} found in select with more than one referenced table/view.".format(
+                            r.raw
+                        ),
+                    )
+                )
+
+        return violation_buff or None
